one more bit

+++ demos/src/lib.rs
@@ -88,19 +88,19 @@ impl Julia {
 
         let mut frag_color = 0u32;
 
-        let ray_direction_x = ray_direction_x.abs();
-        let ray_direction_y = ray_direction_y.abs();
+        let ray_direction_x = ray_direction_x.abs() as u32;
+        let ray_direction_y = ray_direction_y.abs() as u32;
 
-        let mut p_x: i32 = 0;
-        let mut p_y: i32 = 0;
+        let mut p_x: u32 = 0;
+        let mut p_y: u32 = 0;
         let mut p_z: i32 = translate_z;
 
         for _ in 0..ITER_MAX {
-            let index = ((p_x >> (Q+1-7)) * 128 * 128 + (p_y >> (Q+1-7)) * 128 + (p_z >> (Q+1-7))) as usize;
-            let lookup_result = LOOKUP_TABLE[index];
+            let index = (p_x >> (Q+1-7)) * 128 * 128 + (p_y >> (Q+1-7)) * 128 + (p_z as u32 >> (Q+1-7));
+            let lookup_result = LOOKUP_TABLE[index as usize];
             let distance = ((lookup_result >> 24) as u32) << (Q-8);
-            p_x += (ray_direction_x * distance as i32) >> (Q+3);
-            p_y += (ray_direction_y * distance as i32) >> (Q+3);
+            p_x += (ray_direction_x * distance) >> (Q+3);
+            p_y += (ray_direction_y * distance) >> (Q+3);
             p_z += (ray_direction_z * distance as i32) >> (Q+3);
             p_x = p_x & ((2<<Q) - 1);
             p_y = p_y & ((2<<Q) - 1);
